Re-add proper BMA and WS2P API public accesses
BMA (clients), WS2P (server) APIs were simply
not longer publicly accessible

Porting to v2 packaging did not took care of
maintaining this important feature working

Default permission to "/" webadmin is "admins",
if the users choses the default value,
no issue the webadmin is completely not accessible

In case `visitors` is selected, `/` is accessible to visitors,
but its API path `/webmin` is not. Which is hopefully safe.
Just assets are accessible, but no informaton nor command can be
received or sent about data of the instance

It has to be this way, since the CI wants to be able to choose
`visitors` for its tests, probably for beeing able to check that
the app is accessible via its path.